set up bull taskQueues
